add multipe model test
