+++ tests/model_setup/test_init.py
@@ -359,47 +359,3 @@ class TestInitErrors(test.SimpleTestCase):
                     },
                 }
             )
-
-    async def test_init_duplicate_model_name(self):
-        with self.assertRaisesRegex(
-            ConfigurationError, "The model of the Author cannot be created twice."
-        ):
-            await Tortoise.init(
-                {
-                    "connections": {
-                        "default": {
-                            "engine": "tortoise.backends.sqlite",
-                            "credentials": {"file_path": ":memory:"},
-                        }
-                    },
-                    "apps": {
-                        "models": {
-                            "models": ["tests.testmodels", "tests.testmodels2"],
-                            "default_connection": "default",
-                        }
-                    },
-                }
-            )
-        with self.assertRaisesRegex(
-            ConfigurationError, "The model of the Author cannot be created twice."
-        ):
-            await Tortoise.init(
-                {
-                    "connections": {
-                        "default": {
-                            "engine": "tortoise.backends.sqlite",
-                            "credentials": {"file_path": ":memory:"},
-                        }
-                    },
-                    "apps": {
-                        "models": {
-                            "models": ["tests.testmodels"],
-                            "default_connection": "default",
-                        },
-                        "model2": {
-                            "models": ["tests.testmodels2"],
-                            "default_connection": "default",
-                        },
-                    },
-                }
-            )
